Install HAML (read this)
This commit adds the Jekyll Haml plugin[1]. This allows us to use Haml, which
is a simplified version of HTML (less typing 🎉).

Basics of Haml here[2] but do what I did and figure it out from the code, you
won't regret it.

To make it run, you'll need to run:
1. $ gem install jekyll-haml
2. $ jekyll build
3. $ jekyll serve

[1] https://github.com/samvincent/jekyll-haml
[2] http://haml.info/