WIP Create Hero
The hero is a basic module that adds a background image and a highlighted
header. Use modifier classes (classes ending in --word) to make variations of
it for different header positions and/or background images.

In those modifier classes, only add rules that differ and add the modifier
class as a second class.

E.g:
Haml: .m-hero.m-hero--home
HTML: <div class="m-hero m-hero--home"></div>